==> notebooks/00_data_slice_merge.ipynb <==
from pathlib import Path 
import os, dotenv
dotenv.load_dotenv()
os.chdir(Path(os.getenv("PYTHONPATH")).expanduser())

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import json
from src.preprocessing.quality_process import compute_file_metrics

# Source code location
# /Users/josh/SecurityAnalytics/development/cryptol
# /Users/josh/SecurityAnalytics/development/cryptol-specs
# /Users/josh/SecurityAnalytics/development/saw-script
NEW_VERSION = "1.2s"
VARIANTION = "nocomments"
# --- dataset ---
OLD_DATA = f"verified_{VARIANTION}.jsonl"  # e.g., verified_nocomments
# --- paths ---
df = pd.read_json(f"data/training_datasets/{OLD_DATA}", lines=True)
df.head()


with open("syntax_verification_failures.txt", "r") as f:
    files_to_remove = {line.strip() for line in f if line.strip()}

# 2. Drop rows where df["filename"] is in that list
df_filtered = df[~df["filename"].isin(files_to_remove)].reset_index(drop=True).copy()

# optional: if you want to modify df in-place
df = df_filtered

df.info()

df.to_json(f"data/training_datasets/verified_{VARIANTION}_{NEW_VERSION}.jsonl", orient='records', lines=True)
