==> notebooks/00_remove_bad_saw_scripts.ipynb <==
import pandas as pd


df = pd.read_json("data/training_datasets/verified_nomods.jsonl", lines=True)
saw_results_df = pd.read_json("data/training_datasets/syntax_check/saw_results.jsonl", lines=True)
saw_docker_df = pd.read_json("saw_results.jsonl", lines=True)

df.head()

saw_docker_df.rename(columns={"ok": "load_ok"}, inplace=True)
saw_docker_df.head()

saw_results_df.head()

merged_df = pd.concat([saw_results_df[saw_results_df["load_ok"] == True], saw_docker_df], ignore_index=True)


print(len(merged_df["filename"].unique()))
merged_df.head()

merged_df.info()
len(merged_df[merged_df["load_ok"] == True])

print(merged_df["filename"].nunique())
saw_bad_df = merged_df[merged_df["load_ok"] == False]
saw_bad_df.info()

load_status = merged_df.set_index("filename")["load_ok"]

# Define a boolean mask for rows to keep
mask = ~(
    (df["filetype"] == "saw") &                        # only saw files
    (df["filename"].isin(load_status.index)) &          # only filenames that exist in df_load
    (df["filename"].map(load_status) == False)          # and that failed load
)

# Apply mask
filtered_df = df[mask].reset_index(drop=True)

filtered_df.info()

filtered_df.head()

merged_df.to_json("data/training_datasets/syntax_check/saw_results.jsonl", lines=True, orient="records")
filtered_df.to_json("data/training_datasets/verified_nomods.jsonl", lines=True, orient="records")
