==> notebooks/02_comment_process.ipynb <==
from pathlib import Path 
import os, dotenv
dotenv.load_dotenv()
os.chdir(Path(os.getenv("PYTHONPATH")).expanduser())

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import json
from src.preprocessing.comment_process import normalize_blanklines, strip_cryptol_comments_all

jsonl_path = "data/training_datasets/verified_nomods.jsonl"

nomod_df = pd.read_json(jsonl_path, lines=True)

nomod_df.head()

nocomment_df = nomod_df.copy()
nocomment_df["content"] = nocomment_df["content"].apply(strip_cryptol_comments_all)
nocomment_df["content"] = nocomment_df["content"].apply(normalize_blanklines)

nocomment_df.head(20)

nocomment_df.tail(20)

nocomment_df["variant"] = "without_comments"
nocomment_df.head()

nocomment_df.to_json("data/training_datasets/verified_nocomments.jsonl", orient="records", lines=True, force_ascii=False)


from src.preprocessing.comment_extractor import extract_strip_cry_comments
# Create Hybrid Dataset
out_path = Path("cache/GPT_comment_decisions_cache.jsonl")
out_path.parent.mkdir(parents=True, exist_ok=True)  # create dirs if missing
DECISION_CACHE_PATH = "cache/GPT_comment_decisions_cache.jsonl"
comment_rows = []
dataset_rows = []
for index, row in nomod_df.iterrows():
    comments, file_record_ = extract_strip_cry_comments(
        filename=row.filename,
        content=row.content,
        llm_model_name="gpt-oss:20b",
        decision_cache_path=DECISION_CACHE_PATH
    )
    comment_rows.extend(comments)
    file_record = {
        "filename": file_record_["filename"],
        "filetype": row.filetype,
        "content": file_record_["content"],
        "variant": "hybrid"
    }
    dataset_rows.append(file_record)

comment_df = pd.DataFrame(comment_rows)
hybrid_df = pd.DataFrame(dataset_rows)



comment_df.head()

hybrid_df.head()

HYBRID_PATH = "data/training_datasets/verified_hybrid.jsonl"
COMMENT_STAT_PATH = "data/GPTcomment_stats.jsonl"

hybrid_df.to_json(HYBRID_PATH, lines=True, orient="records")
comment_df.to_json(COMMENT_STAT_PATH, lines=True, orient="records")
