==> notebooks/04_text_processing.ipynb <==
from pathlib import Path 
import os, dotenv
dotenv.load_dotenv()
os.chdir(Path(os.getenv("PYTHONPATH")).expanduser())

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import os
from dotenv import load_dotenv

from src.preprocessing.text_agent import get_or_process_text

text_df = pd.read_json("data/parsed_text_data.jsonl", lines=True)
text_df.head()

load_dotenv()
OPEN_AI_KEY = os.getenv("OPEN_AI_KEY")

exclusion_list = [
    "cryptol-specs/Primitive/Keyless/Hash/MD5.md",
    "cryptol-specs/Primitive/Symmetric/Cipher/Block/PRINCE.md",
    "cryptol-specs/Primitive/Symmetric/Cipher/Block/TripleDES.md",
    "cryptol-specs/Primitive/Symmetric/Cipher/Block/DES.md",
    "cryptol-specs/Primitive/Symmetric/Cipher/Block/LED.md",
    "cryptol-specs/Primitive/Symmetric/Cipher/Block/PRINCE.md"
    "cryptol-specs/Primitive/Symmetric/Cipher/Authenticated/ChaChaPolyCryptolIETF.md",
    "cryptol-specs/Primitive/Symmetric/Cipher/Authenticated/ChaChaPolyCryptolIETF.md",
    "cryptol-specs/Primitive/Symmetric/Cipher/Authenticated/SIV_rfc5297.md",
    "cryptol-specs/Primitive/Symmetric/Cipher/Stream/ZUC1_6_Tests.md",
    "cryptol-specs/Primitive/Symmetric/Cipher/Stream/Salsa20.md",
    "cryptol-specs/Primitive/Symmetric/Cipher/Stream/ZUC1_6.md",
    "cryptol-specs/Primitive/Symmetric/KDF/HKDF.md",
    "cryptol-specs/Primitive/Symmetric/KDF/HKDF256Tests.md"
]

text_processing = []
i = 0
for _, row in text_df.iterrows():
    i += 1
    raw = row["content"]
    print(f"Processing row {i} / {len(text_df)}: {row['filename']}")
    if row["filename"] in exclusion_list:
        print(f"  Skipping excluded file: {row['filename']}")
        text_processing.append({
            "filename": row["filename"],
            "hash": None,
            "input": row["content"],
            "processed": raw
        })
        continue
    result = get_or_process_text(raw, model="gpt-5.1", key=OPEN_AI_KEY)
    if isinstance(result, tuple) and len(result) == 2:
        hash, processed = result
    else:
        hash, processed = None, result
    text_processing.append({
        "filename": row["filename"],
        "hash": hash,
        "input": row["content"],
        "processed": processed
    })
    row["content"] = processed
text_processing_df = pd.DataFrame(text_processing)
text_processing_df.head()

text_df.head()

text_df.to_json("data/training_datasets/cleaned_text_data.jsonl", lines=True, orient="records")
